implements #2
now everything gets handled as a double, but you can set an integer too.

+++ android-square-progressbar/src/net/yscs/android/square_progressbar/SquareProgressBar.java
@@ -42,7 +42,7 @@ public class SquareProgressBar extends RelativeLayout {
 		imageView.setImageResource(image);
 	}
 
-	public void setProgress(int progress) {
+	public void setProgress(double progress) {
 		bar.setProgress(progress);
 	}
